Cleanup redundant fixtures and move all other hassio addon fixtures to be shared (#124437)
